requirements.txt
pipasgiref==3.3.4
Django==3.2
django-filter==2.4.0
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
